\name{turnModels-package}
\alias{turnModels-package}
\alias{turnModels}
\docType{package}
\title{
\packageTitle{turnModels}
}
\description{
\packageDescription{turnModels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{turnModels}
\packageIndices{turnModels}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{turnModels}

Maintainer: \packageMaintainer{turnModels}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
